split implementation of serial nmea component from its construction
